requirements.txt
fastapi~=0.115.12
uvicorn[standard]~=0.34.2
pydantic~=2.11.4
python-multipart~=0.0.20
sqlalchemy~=2.0.41
psycopg2-binary~=2.9.10
asyncpg
python-dotenv       # to manage DB credentials via .env
python-jose~=3.4.0
passlib[bcrypt]~=1.7.4
python-magic-bin    # optional for MIME type checking

h11~=0.16.0
pip~=25.0.1
rsa~=4.9.1
pyasn1~=0.4.8
ecdsa~=0.19.1
sniffio~=1.3.1
typing_extensions~=4.13.2
six~=1.17.0
click~=8.2.1
starlette~=0.46.2
pydantic_core~=2.33.2
anyio~=4.9.0
greenlet~=3.2.2
annotated-types~=0.7.0
sqlmodel~=0.0.24